back/requirements.txt
django==3.2.19
djangorestframework==3.14.0
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.2.2
pillow==10.2.0
django-cors-headers
django-oauth-toolkit
django-oauth-toolkit
pyotp
qrcode